[fix] Fix Color_ScalarMultiply
The type of the scalar was set to int, which is wrong. The test did not
catch this as the test used a whole number for the scalar. The correct
type was float.

+++ src/Color.c
@@ -24,7 +24,7 @@ Color Color_Subtract(Color color1, Color color2)
 	return Color_Create(red, green, blue);
 }
 
-Color Color_ScalarMultiply(Color color, int scalar)
+Color Color_ScalarMultiply(Color color, float scalar)
 {
 	float red, green, blue;
 	red = color.red * scalar;
